Refactoriza vistas en views separadas
Se reorganizaron las vistas relacionadas con mototaxis, conductores y novedades, separando cada funcionalidad en funciones individuales dentro la carpeta views_transmedina.
